[6792] Amended placement attributes to be versioned
